document and refactor config events
