+++ common/addons/config-flora/src/main/java/com/dfsek/terra/addons/flora/FloraAddon.java
@@ -9,8 +9,8 @@ import com.dfsek.terra.api.addon.annotations.Addon;
 import com.dfsek.terra.api.addon.annotations.Author;
 import com.dfsek.terra.api.addon.annotations.Version;
 import com.dfsek.terra.api.event.EventListener;
-import com.dfsek.terra.api.event.events.config.ConfigLoadEvent;
-import com.dfsek.terra.api.event.events.config.ConfigPackPreLoadEvent;
+import com.dfsek.terra.api.event.events.config.ConfigurationLoadEvent;
+import com.dfsek.terra.api.event.events.config.pack.ConfigPackPreLoadEvent;
 import com.dfsek.terra.api.injection.annotations.Inject;
 import com.dfsek.terra.api.registry.exception.DuplicateEntryException;
 import com.dfsek.terra.api.util.seeded.BiomeBuilder;
@@ -43,7 +43,7 @@ public class FloraAddon extends TerraAddon implements EventListener {
                 .applyLoader(BlockLayer.class, BlockLayerTemplate::new);
     }
 
-    public void onBiomeLoad(ConfigLoadEvent event) {
+    public void onBiomeLoad(ConfigurationLoadEvent event) {
         if(BiomeBuilder.class.isAssignableFrom(event.getType().getTypeClass())) {
             flora.put(event.getConfiguration().getID(), event.load(new BiomeFloraTemplate()).getFlora());
         }
