+++ common/addons/config-tree/src/main/java/com/dfsek/terra/addons/tree/TreeAddon.java
@@ -8,8 +8,8 @@ import com.dfsek.terra.api.addon.annotations.Addon;
 import com.dfsek.terra.api.addon.annotations.Author;
 import com.dfsek.terra.api.addon.annotations.Version;
 import com.dfsek.terra.api.event.EventListener;
-import com.dfsek.terra.api.event.events.config.ConfigLoadEvent;
-import com.dfsek.terra.api.event.events.config.ConfigPackPreLoadEvent;
+import com.dfsek.terra.api.event.events.config.ConfigurationLoadEvent;
+import com.dfsek.terra.api.event.events.config.pack.ConfigPackPreLoadEvent;
 import com.dfsek.terra.api.injection.annotations.Inject;
 import com.dfsek.terra.api.registry.exception.DuplicateEntryException;
 import com.dfsek.terra.api.util.seeded.BiomeBuilder;
@@ -41,7 +41,7 @@ public class TreeAddon extends TerraAddon implements EventListener {
         event.getPack().applyLoader(TreeLayer.class, TreeLayerTemplate::new);
     }
 
-    public void onBiomeLoad(ConfigLoadEvent event) {
+    public void onBiomeLoad(ConfigurationLoadEvent event) {
         if(BiomeBuilder.class.isAssignableFrom(event.getType().getTypeClass())) {
             trees.put(event.getConfiguration().getID(), event.load(new BiomeTreeTemplate()).getTrees());
         }
