+++ common/api/src/main/java/com/dfsek/terra/api/event/events/config/ConfigurationLoadEvent.java
@@ -1,22 +1,33 @@
 package com.dfsek.terra.api.event.events.config;
 
-import com.dfsek.tectonic.config.Configuration;
+import com.dfsek.tectonic.abstraction.AbstractConfiguration;
+import com.dfsek.tectonic.config.ConfigTemplate;
 import com.dfsek.terra.api.config.ConfigPack;
-import com.dfsek.terra.api.config.Loader;
+import com.dfsek.terra.api.config.ConfigType;
 import com.dfsek.terra.api.event.events.PackEvent;
 
 import java.util.function.Consumer;
 
+/**
+ * Fired when each individual configuration is loaded.
+ * <p>
+ * Addons should listen to this event if they wish to add
+ * config values to existing {@link ConfigType}s.
+ */
 public class ConfigurationLoadEvent implements PackEvent {
     private final ConfigPack pack;
-    private final Loader loader;
+    private final AbstractConfiguration configuration;
+    private final Consumer<ConfigTemplate> loader;
+    private final ConfigType<?, ?> type;
 
-    private final Consumer<Configuration> consumer;
+    private final Object loaded;
 
-    public ConfigurationLoadEvent(ConfigPack pack, Loader loader, Consumer<Configuration> consumer) {
+    public ConfigurationLoadEvent(ConfigPack pack, AbstractConfiguration configuration, Consumer<ConfigTemplate> loader, ConfigType<?, ?> type, Object loaded) {
         this.pack = pack;
+        this.configuration = configuration;
         this.loader = loader;
-        this.consumer = consumer;
+        this.type = type;
+        this.loaded = loaded;
     }
 
     @Override
@@ -24,11 +35,23 @@ public class ConfigurationLoadEvent implements PackEvent {
         return pack;
     }
 
-    public Loader getLoader() {
-        return loader;
+    public AbstractConfiguration getConfiguration() {
+        return configuration;
     }
 
-    public void register(Configuration config) {
-        consumer.accept(config);
+    public <T extends ConfigTemplate> T load(T template) {
+        loader.accept(template);
+        return template;
+    }
+
+    public ConfigType<?, ?> getType() {
+        return type;
+    }
+
+    @SuppressWarnings("unchecked")
+    public <T> T getLoadedObject(Class<T> clazz) {
+        if(!clazz.isAssignableFrom(type.getTypeClass()))
+            throw new ClassCastException("Cannot assign object from loader of type " + type.getTypeClass().getCanonicalName() + " to class " + clazz.getCanonicalName());
+        return (T) loaded;
     }
 }
